requirements.txt
aiogram 
anthropic 
fastmcp 
python-dotenv
langchain
langchain-mcp-adapters
langchain-nvidia-ai-endpoints
